Resenas por calificacion

+++ PROYECTO_DINAMICO/joystickMX_POM/joystick_persistencia/src/main/java/joystickmx/itson/interfaces/IResenaDAO.java
@@ -25,7 +25,7 @@ public interface IResenaDAO {
 
     List<Resena> buscarPorCliente(Long idCliente) throws PersistenciaException;
 
-    List<Resena> buscarPorCalificacion(Integer calificacion) throws PersistenciaException;
+    List<Resena> buscarPorCalificacion(Float calificacion) throws PersistenciaException;
 
     List<Resena> buscarTodas() throws PersistenciaException;
 }
